Рефакторинг функции determine_ext_user + она сделана методом UserService.
