+++ src/bot/handlers/registration.py
@@ -6,25 +6,8 @@ from src.bot.constants import commands
 from src.bot.keyboards import get_start_keyboard
 from src.bot.services.external_site_user import ExternalSiteUserService
 from src.bot.services.user import UserService
-from src.core.db.models import ExternalSiteUser
 from src.core.depends import Container
 from src.core.logging.utils import logger_decor
-
-
-async def determine_ext_user(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-    ext_user_service: ExternalSiteUserService,
-    user_service: UserService,
-) -> ExternalSiteUser | None:
-    if len(context.args) == 1:
-        return await ext_user_service.get_by_id_hash(context.args[0])
-    else:
-        user = await user_service.get_by_telegram_id(update.effective_user.id)
-        if user and user.external_id is not None:
-            return await ext_user_service.get_by_id(user.external_id)
-
-    return None
 
 
 @logger_decor
@@ -35,11 +18,13 @@ async def start_command(
     ext_user_service: ExternalSiteUserService = Provide[Container.bot_services_container.bot_site_user_service],
     user_service: UserService = Provide[Container.bot_services_container.bot_user_service],
 ):
-    ext_user = await determine_ext_user(update, context, ext_user_service, user_service)
+    telegram_user = update.effective_user
+    ext_user = await user_service.determine_ext_user(
+        context.args[0] if len(context.args) == 1 else None, telegram_user.id, ext_user_service
+    )
     if not ext_user:
         return
 
-    telegram_user = update.effective_user
     user = await user_service.register_user(
         telegram_id=telegram_user.id,
         username=telegram_user.username,
